database for products

+++ knit/db_knit/products.py
@@ -6,7 +6,7 @@ class TblProducts(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100))
     desc = db.Column(db.String(2000))
-    price = db.Column(db.Float(12))
+    price = db.Column(db.Integer)
     category = db.Column(db.String(50))
     enable = db.Column(db.Integer)
     picture = db.Column(db.String(50))
